- show snackbar using dateUtils - move updateReceiver to coroutine in WorldClockFragment - show data range or estimate date in AlarmListAdapter - split code of changing view color by switch in AlarmListAdapter - show icon of timezone/ringtone/vibration/snooze if set in AlarmListAdapter - change text if local timezone and added timezone is same - change way to set default ringtone in AlarmActivity

+++ app/src/main/java/com/simples/j/worldtimealarm/WakeUpActivity.kt
@@ -168,7 +168,7 @@ class WakeUpActivity : AppCompatActivity(), View.OnClickListener {
 
             // if snooze is not set, interaction button will work like dismiss
             selector_layout.startRippleAnimation()
-            if(item.snooze == 0.toLong() || intent.action == AlarmReceiver.ACTION_SNOOZE) {
+            if(item.snooze == 0L || intent.action == AlarmReceiver.ACTION_SNOOZE) {
                 interaction_button.setImageDrawable(getDrawable(R.drawable.ic_action_alarm_off))
             }
         }
